+++ nginx-rs/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "nginx-rs"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+
+[build-dependencies]
+bindgen = "0.64"
